% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_aggregate_cwis.R
\name{aggregate_cwis}
\alias{aggregate_cwis}
\title{Aggregate daily CWIS estimates}
\usage{
aggregate_cwis(results)
}
\arguments{
\item{results}{A list of successful \code{cwis_day_result} objects and/or failed
day records created by \code{\link[=cwis]{cwis()}}.}
}
\value{
A list with daily results, information-share summary, diagnostic
summary, and variance-weight summary.
}
\description{
Aggregate daily CWIS estimates
}
