% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasbrouck_95_price_discovery.R
\name{price_discovery}
\alias{price_discovery}
\title{Price discovery analyses between multiple markets}
\usage{
price_discovery(
  data,
  num_market = NULL,
  price_columns = NULL,
  log_price = TRUE,
  lag_selection = TRUE,
  vecm_max.lag = NULL,
  lag_select_ceritera = "SC",
  vecm_lag = NULL,
  coin_const = FALSE,
  coin_beta = TRUE
)
}
\arguments{
\item{data}{a `data.table` object of market data that contains all market prices.
Each column should represent each market price.}

\item{num_market}{Number of markets in the data.}

\item{price_columns}{A vector that includes all market price column names.}

\item{log_price}{Whether market prices should be transformed to natural logarithms,
the default is `TRUE`.}

\item{lag_selection}{Whether selecting VECM's lag based on information criteria.}

\item{vecm_max.lag}{When `lag_selection=TRUE`, the maximum lag length in the
vector error correction model (VECM).}

\item{lag_select_ceritera}{The information criterion is used to select VECM's lag.}

\item{vecm_lag}{The lag length is used in the VECM estimation when `lag_selection=FALSE`}

\item{coin_const}{Whether to include constant term into long-run equilibrium relationship.}

\item{coin_beta}{Whether the cointegrating beta vector is `[1, -1]`, which
is standard setting in Hasbrouck (1995).}
}
\value{
A `data.table` object that stores component share, information share (IS),
and information leadership share (ILS).
}
\description{
`price_discovery()` is a function to calculate price discovery measures under
Hasbrouck (1995)'s one-security-many-markets setting.
}
\examples{

\dontrun{

test <- price.discovery(data = data,
                        num_market = 5,
                        price_columns = c("mkt1", "mkt2", "mkt3", "mkt4", "mkt5"),
                        log_price = TRUE,
                        vecm_max.lag = 60,
                        lag_selection = TRUE,
                        lag_select_ceritera = "SC",
                        coin_const = TRUE,
                        coin_beta = TRUE)
                        }

}
