% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_realized_variance.R
\name{realized_variance}
\alias{realized_variance}
\title{Estimate realized variance from efficient-price returns}
\usage{
realized_variance(returns)
}
\arguments{
\item{returns}{Numeric efficient-price returns.}
}
\value{
Sum of squared returns.
}
\description{
Estimate realized variance from efficient-price returns
}
