% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Market data resampling}
\usage{
resample(
  mkt_data,
  time_column,
  sample_freq,
  sample_period,
  market_open,
  market_close,
  fill = TRUE
)
}
\arguments{
\item{mkt_data}{A `data.table` object of market data set.}

\item{time_column}{The column name that contains the timestamp, or time variable.}

\item{sample_freq}{The sampling frequency, e.g., second, minute.}

\item{sample_period}{The sampling period, e.g., 1, 5, 10}

\item{market_open}{The market open time that is in the format HH:MM:SS, 24-hour clock}

\item{market_close}{The market close time that is in the format HH:MM:SS, 24-hour clock}

\item{fill}{if there is no observation available at a specific timestamp, whether is should be
filled with the most recent value}
}
\value{
A `data.table` object of resampled market dataset ggplot2
}
\description{
`resample_data()` resamples the original data to other sampling frequency. This
function is based on `aggregatePrice` function from `highfrequency` package.
}
\examples{
\dontrun{

resampled_data <- resample(data,'time', "secs", 1, "00:00:00", "23:59:59")
## Column with timestamp is "time"
## resampled at 1 second level
# 24-hour market, market starts at 00:00:
}
}
