% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_vecm.R
\name{vecm}
\alias{vecm}
\title{Estimate the overlapping-period VECM}
\usage{
vecm(log_prices, K = 10, type = "trace", ecdet = "none", spec = "transitory")
}
\arguments{
\item{log_prices}{A two-column matrix or data frame of log futures and spot
prices.}

\item{K}{Lag order passed to \code{\link[urca:ca.jo]{urca::ca.jo()}}.}

\item{type, ecdet, spec}{Arguments passed to \code{\link[urca:ca.jo]{urca::ca.jo()}}.}
}
\value{
A \code{ca.jo} object.
}
\description{
Estimate the overlapping-period VECM
}
