% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_combine_weights.R
\name{combine_weights}
\alias{combine_weights}
\title{Combine period estimates and calculate variance weights}
\usage{
combine_weights(overlapping_variance, non_overlapping_variance)
}
\arguments{
\item{overlapping_variance}{Named realized variances from overlapping
sessions.}

\item{non_overlapping_variance}{Named realized kernels from single-market
sessions.}
}
\value{
A list containing period variances, variance weights, and the total
overlapping-period weight.
}
\description{
Combine period estimates and calculate variance weights
}
