% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_his.R
\name{his}
\alias{his}
\title{Compute Hasbrouck information-share bounds from a VECM}
\usage{
his(vecm_fit)
}
\arguments{
\item{vecm_fit}{A fitted \code{ca.jo} object.}
}
\value{
A named numeric vector for futures and spot.
}
\description{
Compute Hasbrouck information-share bounds from a VECM
}
