% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_split_trading_days.R
\name{split_trading_days}
\alias{split_trading_days}
\title{Split standardized observations by trading day}
\usage{
split_trading_days(data)
}
\arguments{
\item{data}{Standardized data returned by
\code{\link[=validate_and_standardize_input]{validate_and_standardize_input()}}.}
}
\value{
A named list of daily data frames.
}
\description{
Split standardized observations by trading day
}
